requirements.txt
certifi==2020.12.5
chardet==4.0.0
Django==2.2.4
django-cors-headers==3.7.0
django-environ==0.4.5
django-filter==2.4.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.6.0
gunicorn==19.9.0
idna==2.10
psycopg2==2.8.3
PyJWT==2.0.1
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
urllib3==1.26.3